fix issue with new superuser proc
fixing an issue when new companies are created and the system assigned roles and rights to the new company. the Leasing and Rents conversion tool right was broken in march because the name was changed incorrectly. this has been fixed. also removed adding the new access to help center right from being added as it is not yet ready to be deployed.