Merged PR 66033: Moved RAUL label mapping to product settings table
Removed the helper method for Product -> RAUL label mapping and replaced calls to it with a call to get product settings and filter to a new ProductIcon setting type.

Created a data script to populate all the mapping values previously in code to a product setting entry and cascaded up to a global product configuration.

Related work items: #676347