Merged PR 203823: removing properties that dont exist
removing properties that show as no product. these properties were in UDM for the product, but when we tried to find them in the product property api result, they could not be found, so it was decided to just remove them from the audit page since they could just be mapped incorrectly or no longer used properties.

Related work items: #1526262