Merged PR 349084: User Story 2375427: Create new product setting 'FriendlyURLProductName'
#2375427

Update post-deployment scripts and add new product setting

Removed reference to `2255379-consume-demo-orgtype.sql` in `Identity.sqlproj` and added `2375427-product-add-setting-friendlyurlproductname.sql`. The new script checks for the existence of the 'FriendlyURLProductName' setting in the `Enterprise.ProductSettingType` table and inserts it if it does not exist. Updated `Script.PostDeployment.sql` to include the new script and commented out the old reference.